taille int64

+++ GestionBoutiqueFlorale/Client.cs
@@ -12,7 +12,7 @@ namespace GestionBoutiqueFlorale
         [JsonIgnore]
         public List<Commande> Commandes { get; set; }
 
-        public Client(string nom, string prenom, int telephone, string adresse)
+        public Client(string nom, string prenom, long telephone, string adresse)
             : base(nom, prenom, telephone, adresse)
         {
              Commandes= new List<Commande>();
